Move glsl support files to avoid conflict
